Expression parser had a bug where returning null values would leak other data.
