+++ expression-js/Cargo.toml
@@ -11,4 +11,4 @@ expression = { path = ".." }
 wasm-bindgen = { version = "0.2.95" }
 serde = { version = "1.0.210", features = ["derive"] }
 js-sys = "0.3.72"
-web-sys = "0.3.72"
+web-sys = { version = "0.3.72", features = ["console"] }
